pyproject.toml
[project]
name = "epub-to-mdbook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "html2text>=2024.2.26",
    "python-slugify>=8.0.4",
]
